refactor: separate streaming responsibilities
Move transport, timeout, parsing, relay, and audit logic into focused files so each path can be understood and tested independently.
