+++ internal/mcp/client.go
@@ -720,1121 +720,6 @@ func (c *Client) invokeHTTP(ctx context.Context, upstream Upstream, tool string,
 	return invoke, nil
 }
 
-// StreamEvent is one intermediate upstream JSON-RPC message, independent of
-// whether HTTP SSE or stdio carried it. It is either a notification (progress,
-// logging, or another server-to-client notification) or, more rarely, a
-// server-to-client request.
-type StreamEvent struct {
-	// Data is one raw JSON-RPC message. HTTP SSE joins the event's data lines
-	// with newlines; stdio removes its newline framing.
-	Data []byte
-	// ServerRequest is true when Data is a JSON-RPC request from the
-	// upstream (has both id and method) rather than a notification. Atryum
-	// does not broker server-initiated requests (sampling, elicitation,
-	// roots); these are surfaced to the sink for audit only, never relayed
-	// to the agent.
-	ServerRequest bool
-}
-
-// StreamSink receives intermediate upstream messages live, as InvokeStream
-// reads them, so a caller can relay them onward (or just audit them) before
-// the terminal response exists. Its methods run synchronously on the same
-// goroutine as the InvokeStream call — there is no concurrent access to the
-// sink, and no need for the sink to synchronize internally on that account.
-type StreamSink interface {
-	// StreamStarted fires at most once. HTTP SSE calls it before the first
-	// event or terminal response; stdio calls it only before the first
-	// intermediate event. A silently retried attempt never calls it.
-	StreamStarted()
-	// Event delivers one intermediate (non-terminal) message. A returned
-	// error aborts the stream: InvokeStream stops reading and returns that
-	// error to its caller.
-	Event(evt StreamEvent) error
-}
-
-// StreamOptions bounds InvokeStream's setup and response-reading phases. A
-// zero-valued field disables that particular bound.
-type StreamOptions struct {
-	// HeaderTimeout bounds setup before tool response reading begins: HTTP
-	// session initialization and response headers, or the stdio initialize
-	// handshake. Zero leaves setup bounded only by ctx's deadline, if any.
-	HeaderTimeout time.Duration
-	// IdleTimeout bounds response-reading inactivity. Streaming transports
-	// reset it when upstream activity arrives, including events routed over
-	// the shared standalone HTTP stream. For a plain HTTP JSON response it
-	// bounds the complete body read. Zero disables the check.
-	IdleTimeout time.Duration
-	// MaxDuration bounds the complete response-reading phase after HTTP
-	// headers or the stdio handshake. Zero disables the check.
-	MaxDuration time.Duration
-}
-
-type rpcMessageKind int
-
-const (
-	rpcMessageUnknown rpcMessageKind = iota
-	rpcMessageTerminalResponse
-	rpcMessageNotification
-	rpcMessageServerRequest
-)
-
-// classifyRPCMessage identifies one already-parsed JSON-RPC message for the
-// streaming relay: the terminal response to our request (matches
-// expectedID, or is the null-id error the JSON-RPC spec uses when a server
-// can't identify which request an error belongs to), a notification (no id),
-// a server-to-client request (id and method, no result/error), or unknown
-// (e.g. a response to some other id — not ours to interpret). Transport-
-// neutral: used for both the HTTP SSE relay (relaySSEToolCall) and the
-// stdio relay (relayStdioToolCall), and for stdio's non-streaming readRPC,
-// since a JSON-RPC message's shape doesn't depend on how it was framed on
-// the wire.
-func classifyRPCMessage(payload []byte, expectedID json.RawMessage) rpcMessageKind {
-	var message map[string]json.RawMessage
-	if err := json.Unmarshal(payload, &message); err != nil {
-		return rpcMessageUnknown
-	}
-	id, hasID := message["id"]
-	_, hasMethod := message["method"]
-	_, hasResult := message["result"]
-	_, hasError := message["error"]
-
-	if hasID && (hasResult || hasError) {
-		if hasError && jsonRawIsNull(id) {
-			return rpcMessageTerminalResponse
-		}
-		if jsonRPCIDsMatch(id, expectedID) {
-			return rpcMessageTerminalResponse
-		}
-		return rpcMessageUnknown
-	}
-	if hasID && hasMethod {
-		return rpcMessageServerRequest
-	}
-	if !hasID && hasMethod {
-		return rpcMessageNotification
-	}
-	return rpcMessageUnknown
-}
-
-// callTimeoutGuard implements InvokeStream's setup/idle/max-duration timeout
-// scheme by canceling one shared context. The setup timer covers HTTP response
-// headers or the stdio initialize handshake. After setup, callers replace it
-// with idle and maximum-duration timers for response reading.
-type callTimeoutGuard struct {
-	ctx    context.Context
-	cancel context.CancelFunc
-
-	// mu guards trippedWhy, stopped, the timer fields, and idleTimeout.
-	// The timer fields need it because a time.AfterFunc callback starts
-	// its clock before the assignment of the returned *Timer completes:
-	// checkIdle (running on the timer's goroutine) could otherwise read
-	// g.idleTimer before/while armBodyTimeouts writes it — a data race by
-	// the memory model even if the window is nanoseconds in practice.
-	mu          sync.Mutex
-	trippedWhy  string
-	stopped     bool
-	headerTimer *time.Timer
-	idleTimer   *time.Timer
-	maxTimer    *time.Timer
-	idleTimeout time.Duration
-
-	// lastActivity (unix nanoseconds) is updated by resetIdle and read by
-	// checkIdle. It exists so the idle timer's firing can be verified
-	// rather than trusted outright — see checkIdle. Atomic, not mu-guarded:
-	// resetIdle runs once per relayed event on the hot path and must not
-	// contend with the timer goroutine.
-	lastActivity atomic.Int64
-}
-
-func newCallTimeoutGuard(parent context.Context) *callTimeoutGuard {
-	ctx, cancel := context.WithCancel(parent)
-	return &callTimeoutGuard{ctx: ctx, cancel: cancel}
-}
-
-func (g *callTimeoutGuard) trip(why string) {
-	g.mu.Lock()
-	if g.trippedWhy == "" {
-		g.trippedWhy = why
-	}
-	g.mu.Unlock()
-	g.cancel()
-}
-
-func (g *callTimeoutGuard) armHeaderTimeout(d time.Duration) {
-	if d <= 0 {
-		return
-	}
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	if g.stopped {
-		return
-	}
-	g.headerTimer = time.AfterFunc(d, func() { g.trip("timed out waiting for upstream response headers") })
-}
-
-func (g *callTimeoutGuard) disarmHeaderTimeout() {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	if g.headerTimer != nil {
-		g.headerTimer.Stop()
-	}
-}
-
-func (g *callTimeoutGuard) armBodyTimeouts(idle, max time.Duration) {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	if g.stopped {
-		return
-	}
-	g.idleTimeout = idle
-	if idle > 0 {
-		g.lastActivity.Store(time.Now().UnixNano())
-		g.idleTimer = time.AfterFunc(idle, g.checkIdle)
-	}
-	if max > 0 {
-		g.maxTimer = time.AfterFunc(max, func() { g.trip("max stream duration exceeded") })
-	}
-}
-
-// checkIdle is the idle timer's callback. It does not trust "the timer
-// fired" to mean "genuinely idle": time.Timer.Reset called concurrently
-// with a timer's own firing is explicitly documented as racy (the AfterFunc
-// callback may already be running by the time Reset takes effect), so
-// resetIdle deliberately never calls Reset at all — it only records the
-// latest activity timestamp. checkIdle re-derives the real elapsed time
-// from that timestamp and either trips (elapsed genuinely exceeds the
-// bound) or reschedules for the remaining time (an event arrived
-// concurrently with this firing). This makes the idle bound correct
-// regardless of how resetIdle and the timer callback interleave. The
-// stopped check makes a firing that lost the race with stop() a no-op
-// instead of re-arming a timer the guard's owner believes is dead.
-func (g *callTimeoutGuard) checkIdle() {
-	g.mu.Lock()
-	if g.stopped {
-		g.mu.Unlock()
-		return
-	}
-	idleTimeout := g.idleTimeout
-	elapsed := time.Duration(time.Now().UnixNano() - g.lastActivity.Load())
-	if elapsed < idleTimeout {
-		if g.idleTimer != nil {
-			g.idleTimer.Reset(idleTimeout - elapsed)
-		}
-		g.mu.Unlock()
-		return
-	}
-	g.mu.Unlock()
-	// trip acquires g.mu itself; called outside the lock.
-	g.trip("idle timeout waiting for the next stream event")
-}
-
-func (g *callTimeoutGuard) resetIdle() {
-	g.lastActivity.Store(time.Now().UnixNano())
-}
-
-// stop is idempotent: the guard's owners defer it both at guard creation
-// (covering early-error returns) and inside subprocess-cleanup defers that
-// must cancel the context before waiting on the process.
-func (g *callTimeoutGuard) stop() {
-	g.mu.Lock()
-	g.stopped = true
-	if g.headerTimer != nil {
-		g.headerTimer.Stop()
-	}
-	if g.idleTimer != nil {
-		g.idleTimer.Stop()
-	}
-	if g.maxTimer != nil {
-		g.maxTimer.Stop()
-	}
-	g.mu.Unlock()
-	g.cancel()
-}
-
-func (g *callTimeoutGuard) reason() string {
-	g.mu.Lock()
-	defer g.mu.Unlock()
-	return g.trippedWhy
-}
-
-// streamCallOutcome is the result of one attempt to send a streaming
-// tools/call request. missingSession mirrors doHTTPEnvelope's
-// SessionExpired signal so invokeHTTPStream can apply the same
-// reinitialize-and-retry-once policy invokeHTTP uses — but only when
-// eventsRelayed is 0: once anything has reached the sink, the downstream
-// has already seen stream bytes, so retrying would relay a second copy of
-// everything. In that case the caller fails instead of retrying.
-type streamCallOutcome struct {
-	invoke         InvokeResult
-	missingSession bool
-	sessionID      string
-	eventsRelayed  int
-}
-
-// doHTTPToolCallStream sends one tools/call request and either reads a
-// plain JSON body (mapped exactly like the buffered path) or, for an SSE
-// response, relays intermediate events to sink live and returns once the
-// terminal response is read.
-func (c *Client) doHTTPToolCallStream(ctx context.Context, upstream Upstream, body []byte, sink StreamSink, progressCh <-chan StreamEvent, opts StreamOptions) (streamCallOutcome, error) {
-	guard := newCallTimeoutGuard(ctx)
-	defer guard.stop()
-	guard.armHeaderTimeout(opts.HeaderTimeout)
-
-	h, err := c.doHTTPEnvelopeHeaders(guard.ctx, upstream, body, DefaultMCPProtocolVersion, true)
-	guard.disarmHeaderTimeout()
-	if err != nil {
-		if reason := guard.reason(); reason != "" {
-			return streamCallOutcome{}, fmt.Errorf("upstream %q: %s: %w", upstream.Name, reason, ErrStreamTimeout)
-		}
-		return streamCallOutcome{}, err
-	}
-	resp := h.resp
-
-	// Armed as soon as headers are back, before any body is read — the
-	// header timeout only ever bounded waiting for headers, so every
-	// body-reading branch below (including the two early returns, not just
-	// the SSE relay) needs its own bound. Without this, a slow/hanging body
-	// on a 404-session-expired or plain-JSON response during a streaming
-	// call would be unbounded: the per-call http.Client.Timeout that would
-	// normally catch this is deliberately skipped in streaming mode (see
-	// doHTTPEnvelopeRaw's streaming param).
-	guard.armBodyTimeouts(opts.IdleTimeout, opts.MaxDuration)
-
-	if h.sessionExpired {
-		defer resp.Body.Close()
-		_, _ = io.Copy(io.Discard, resp.Body)
-		return streamCallOutcome{missingSession: true, sessionID: h.sessionID}, nil
-	}
-
-	if !strings.Contains(strings.ToLower(h.contentType), "text/event-stream") {
-		defer resp.Body.Close()
-		bodyBytes, err := io.ReadAll(resp.Body)
-		if err != nil {
-			if reason := guard.reason(); reason != "" {
-				return streamCallOutcome{}, fmt.Errorf("upstream %q: %s: %w", upstream.Name, reason, ErrStreamTimeout)
-			}
-			return streamCallOutcome{}, err
-		}
-		forward := ForwardResult{StatusCode: resp.StatusCode, Body: bodyBytes, ContentType: h.contentType, ProtocolVersion: h.protocolVersion, SessionID: h.sessionID}
-		invoke, missingSession, err := toolCallResultFromForward(forward)
-		if err != nil {
-			return streamCallOutcome{}, err
-		}
-		return streamCallOutcome{invoke: invoke, missingSession: missingSession, sessionID: h.sessionID}, nil
-	}
-
-	// relaySSEToolCall owns resp.Body because it may replace this response
-	// with one or more resumed GET streams before the terminal response.
-	return c.relaySSEToolCall(resp, sink, progressCh, guard, upstream, h.sessionID)
-}
-
-// postStreamMsg is one message pumped from a tools/call POST response by
-// postStreamPump: either a data-bearing JSON-RPC payload (data != nil), or a
-// terminal error ending the stream (err != nil).
-type postStreamMsg struct {
-	data []byte
-	err  error
-}
-
-// postStreamPump owns the tools/call POST response's read loop — including
-// SSE resumption — on its own goroutine, feeding relaySSEToolCall with only
-// the data-bearing JSON-RPC payloads (or a final error) through msgs. This
-// lets relaySSEToolCall select between this stream and a per-call
-// standalone-stream channel (progressCh) without either blocking the
-// other, so a call is only ever done reading (and only ever returns to its
-// caller) once both are accounted for — see progressWaiter for why that
-// matters.
-type postStreamPump struct {
-	msgs chan postStreamMsg
-
-	mu       sync.Mutex
-	current  *http.Response
-	stopped  bool
-	stopOnce sync.Once
-	done     chan struct{}
-}
-
-func newPostStreamPump(c *Client, guard *callTimeoutGuard, upstream Upstream, resp *http.Response) *postStreamPump {
-	p := &postStreamPump{msgs: make(chan postStreamMsg), current: resp, done: make(chan struct{})}
-	go p.run(c, guard, upstream)
-	return p
-}
-
-// stop closes the currently-active response body, if any — causing a
-// blocked Read to return promptly — and marks the pump stopped so it exits
-// instead of trying to resume. Safe to call more than once; only the first
-// call has any effect. Always safe to call even if the pump has already
-// finished on its own.
-func (p *postStreamPump) stop() {
-	p.stopOnce.Do(func() {
-		p.mu.Lock()
-		p.stopped = true
-		cur := p.current
-		p.mu.Unlock()
-		close(p.done)
-		if cur != nil {
-			_ = cur.Body.Close()
-		}
-	})
-}
-
-// setCurrent installs resp as the response the pump is currently reading
-// from (after a resume). Returns false — and leaves resp to the caller to
-// close — if stop was already called, so a resume racing a stop can't
-// resurrect a pump that's supposed to be shutting down.
-func (p *postStreamPump) setCurrent(resp *http.Response) bool {
-	p.mu.Lock()
-	defer p.mu.Unlock()
-	if p.stopped {
-		return false
-	}
-	p.current = resp
-	return true
-}
-
-// send delivers msg, or exits early if stop is called while blocked trying
-// to (msgs is unbuffered: without this, a caller that stops reading msgs
-// after its own terminal response — see relaySSEToolCall — would otherwise
-// leave this goroutine permanently blocked on a send nobody will ever
-// receive).
-func (p *postStreamPump) send(msg postStreamMsg) {
-	select {
-	case p.msgs <- msg:
-	case <-p.done:
-	}
-}
-
-func (p *postStreamPump) run(c *Client, guard *callTimeoutGuard, upstream Upstream) {
-	defer close(p.msgs)
-	reader := newSSEEventReader(p.current.Body)
-	lastEventID := ""
-	retryDelay := time.Duration(0)
-	// resumedFrom holds, after a resume, the cursor id the Last-Event-ID
-	// header carried. Replay semantics are exclusive of the cursor, but the
-	// classic server off-by-one replays it inclusively — without this guard
-	// the cursor event's data would be relayed to the agent a second time.
-	// The guard window closes at the first event bearing any other id, so a
-	// server legitimately reusing the id much later is unaffected.
-	resumedFrom := ""
-	for {
-		evt, err := reader.NextEvent()
-		if err != nil {
-			p.mu.Lock()
-			stopped := p.stopped
-			p.mu.Unlock()
-			if stopped {
-				return
-			}
-			if reason := guard.reason(); reason != "" {
-				p.send(postStreamMsg{err: fmt.Errorf("upstream %q %s: %w", upstream.Name, reason, ErrStreamTimeout)})
-				return
-			}
-			if err != io.EOF {
-				p.send(postStreamMsg{err: err})
-				return
-			}
-			if lastEventID == "" {
-				p.send(postStreamMsg{err: fmt.Errorf("upstream %q closed the stream without a JSON-RPC response or resumable event id", upstream.Name)})
-				return
-			}
-			if err := waitForSSEReconnect(guard.ctx, retryDelay); err != nil {
-				if reason := guard.reason(); reason != "" {
-					p.send(postStreamMsg{err: fmt.Errorf("upstream %q %s while waiting to resume: %w", upstream.Name, reason, ErrStreamTimeout)})
-					return
-				}
-				p.send(postStreamMsg{err: err})
-				return
-			}
-			resumed, err := c.resumeSSEStream(guard.ctx, upstream, lastEventID)
-			if err != nil {
-				if reason := guard.reason(); reason != "" {
-					p.send(postStreamMsg{err: fmt.Errorf("upstream %q %s while resuming: %w", upstream.Name, reason, ErrStreamTimeout)})
-					return
-				}
-				p.send(postStreamMsg{err: err})
-				return
-			}
-			if !p.setCurrent(resumed) {
-				_ = resumed.Body.Close()
-				return
-			}
-			reader = newSSEEventReader(resumed.Body)
-			resumedFrom = lastEventID
-			continue
-		}
-		guard.resetIdle()
-		if evt.HasRetry {
-			retryDelay = evt.Retry
-		}
-		if evt.HasID {
-			if resumedFrom != "" && evt.ID == resumedFrom {
-				// Inclusive replay of the cursor event we already relayed
-				// before the disconnect: keep the bookkeeping, skip the data.
-				lastEventID = evt.ID
-				continue
-			}
-			resumedFrom = ""
-			lastEventID = evt.ID
-		}
-		if !evt.HasData {
-			continue
-		}
-		p.send(postStreamMsg{data: evt.Data})
-	}
-}
-
-// relaySSEToolCall reads resp's SSE body incrementally (via postStreamPump,
-// on a dedicated goroutine), relaying every intermediate (non-terminal)
-// message to sink as it arrives, and returns once the terminal JSON-RPC
-// response for id "1" is read. It also drains progressCh — standalone-
-// stream notifications matched to this call (see progressWaiter) — via the
-// same select loop, so exactly one goroutine ever calls sink.Event for a
-// given call. resp.Body is not closed here directly; postStreamPump owns
-// that (including across resumes, which replace it with a new response).
-// sessionID is the session this attempt was sent under; it is always
-// stamped onto the returned outcome (even a missing-session terminal
-// response) so a caller retry can identify and clear the right session —
-// mirroring doHTTPEnvelope's ForwardResult.SessionID contract.
-//
-// sink.StreamStarted fires lazily, right before the first thing is actually
-// delivered — not simply because the response's Content-Type was SSE. This
-// matters for the missing-session retry: if the very first (and only)
-// message is a missing-session terminal error, the whole attempt is
-// discarded and silently retried (see invokeHTTPStream), so the sink must
-// never have been told a stream started for it. Once a real event has been
-// relayed, or the terminal response is anything other than a
-// zero-events missing-session error, the attempt is the one that counts and
-// StreamStarted fires exactly once for it.
-func (c *Client) relaySSEToolCall(resp *http.Response, sink StreamSink, progressCh <-chan StreamEvent, guard *callTimeoutGuard, upstream Upstream, sessionID string) (streamCallOutcome, error) {
-	expectedID := json.RawMessage([]byte("1"))
-	statusCode := resp.StatusCode
-	relayed := 0
-	started := false
-	ensureStarted := func() {
-		if !started {
-			started = true
-			sink.StreamStarted()
-		}
-	}
-	deliver := func(evt StreamEvent) error {
-		guard.resetIdle()
-		ensureStarted()
-		relayed++
-		return sink.Event(evt)
-	}
-
-	pump := newPostStreamPump(c, guard, upstream, resp)
-	defer pump.stop()
-
-	for {
-		select {
-		case evt, ok := <-progressCh:
-			if !ok {
-				// Never actually closed (its registration outlives this
-				// call — see invokeHTTPStream's grace period), but nil this
-				// out defensively so a closed channel can't busy-loop.
-				progressCh = nil
-				continue
-			}
-			if err := deliver(evt); err != nil {
-				return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, err
-			}
-		case msg, ok := <-pump.msgs:
-			if !ok {
-				return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, fmt.Errorf("upstream %q: stream ended unexpectedly", upstream.Name)
-			}
-			if msg.err != nil {
-				return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, msg.err
-			}
-			payload := msg.data
-			switch classifyRPCMessage(payload, expectedID) {
-			case rpcMessageTerminalResponse:
-				var rpcResp rpcResponse
-				if err := json.Unmarshal(payload, &rpcResp); err != nil {
-					return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, err
-				}
-				invoke, missingSession := toolCallResultFromRPCResponse(rpcResp, statusCode)
-				if progressCh != nil {
-					// See terminalSettleWindow: give a notification already in
-					// flight on the standalone stream a brief, bounded chance
-					// to arrive before finalizing.
-					settle := time.NewTimer(terminalSettleWindow)
-				settleLoop:
-					for {
-						select {
-						case evt, ok := <-progressCh:
-							if !ok {
-								break settleLoop
-							}
-							if err := deliver(evt); err != nil {
-								settle.Stop()
-								return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, err
-							}
-							if !settle.Stop() {
-								<-settle.C
-							}
-							settle.Reset(terminalSettleWindow)
-						case <-settle.C:
-							break settleLoop
-						}
-					}
-				}
-				if !(missingSession && relayed == 0) {
-					ensureStarted()
-				}
-				return streamCallOutcome{invoke: invoke, missingSession: missingSession, eventsRelayed: relayed, sessionID: sessionID}, nil
-			case rpcMessageServerRequest:
-				if err := deliver(StreamEvent{Data: payload, ServerRequest: true}); err != nil {
-					return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, err
-				}
-			case rpcMessageNotification:
-				if err := deliver(StreamEvent{Data: payload}); err != nil {
-					return streamCallOutcome{eventsRelayed: relayed, sessionID: sessionID}, err
-				}
-			default:
-				// Unrecognized payload shape (e.g. a response to some other id).
-				// Not ours to interpret; ignore and keep reading.
-			}
-		}
-	}
-}
-
-func waitForSSEReconnect(ctx context.Context, delay time.Duration) error {
-	if delay <= 0 {
-		return nil
-	}
-	timer := time.NewTimer(delay)
-	defer timer.Stop()
-	select {
-	case <-timer.C:
-		return nil
-	case <-ctx.Done():
-		return ctx.Err()
-	}
-}
-
-// progressWaiter is one streaming call's registration with a
-// standaloneStream: a notification matching wireToken has its raw payload
-// sent to events. events is buffered and drained only by
-// relaySSEToolCall's own goroutine (via select, alongside that call's own
-// POST-response reads — see postStreamPump) rather than delivered directly
-// to the sink from the shared reader goroutine that owns routeStandaloneEvent.
-// That indirection is what makes it safe: without it, a notification for
-// this call and this call's own terminal response race across two
-// independent goroutines with no ordering guarantee, and a delivery
-// attempt could land after the call has already returned to its caller —
-// which, at the HTTP handler layer, may already have written the terminal
-// SSE frame and returned, making a later write to the same
-// http.ResponseWriter unsafe.
-type progressWaiter struct {
-	events chan StreamEvent
-}
-
-// standaloneWaiterEventBuffer bounds progressWaiter.events. Sized generously
-// relative to realistic progress-update rates: a full buffer means the
-// receiving call's own goroutine isn't draining it (already finished, or
-// deep in its own resume/retry handling), in which case routeStandaloneEvent
-// drops the event rather than blocking — a shared reader goroutine also
-// serving other concurrent calls must never block on one slow receiver.
-const standaloneWaiterEventBuffer = 32
-
-// standaloneStream manages one shared "standalone" SSE GET connection per
-// upstream — the channel the MCP Streamable HTTP transport defines for
-// server-initiated messages that aren't tied to any specific request.
-//
-// This exists because the reference MCP Python SDK's Context.report_progress
-// does not attribute its notification to the request that triggered it (it
-// calls send_progress_notification without related_request_id), so the
-// server's message router sends it to this standalone stream, never to the
-// tools/call POST response body that relaySSEToolCall reads. Without this,
-// Atryum cannot see those notifications at all.
-//
-// Atryum multiplexes every downstream caller of a given upstream onto one
-// shared session, so this stream is refcounted across concurrent streaming
-// calls rather than opened per call: acquireStandaloneStream starts the
-// connection for the first waiter and releaseStandaloneStream tears it down
-// once the last waiter is gone. It deliberately does not implement
-// Last-Event-ID resumption (unlike relaySSEToolCall's per-call stream): if
-// the connection drops mid-flight, any calls still waiting on it simply stop
-// receiving standalone-routed notifications until the next acquire cycle
-// reopens it — an accepted limitation, not a correctness hazard, since the
-// call's own terminal response still arrives normally on its POST stream.
-//
-// standaloneWaiterGracePeriod bounds how long a call's progressWaiter
-// lingers in the waiters map after the call itself has completed, before
-// invokeHTTPStream's deferred cleanup actually removes it. See that cleanup
-// for why immediate removal is unsafe.
-const standaloneWaiterGracePeriod = 2 * time.Second
-
-// terminalSettleWindow bounds how long relaySSEToolCall waits, once it has
-// read this call's terminal response, for anything further to arrive on
-// progressCh before finalizing — reset each time something does arrive, so
-// a burst of trailing notifications is fully drained rather than cut off
-// after one. This call's own POST response and the shared standalone
-// stream are independent connections read by independent goroutines: even
-// with progressWaiter's channel already holding a pending notification by
-// the time the terminal is read, Go's select has no rule preferring one
-// ready case over another, so without this window a notification that
-// arrived at essentially the same instant as the terminal could be skipped
-// — not because it never arrived, but because select happened not to pick
-// it up first.
-const terminalSettleWindow = 25 * time.Millisecond
-
-type standaloneStream struct {
-	mu       sync.Mutex
-	refCount int
-	cancel   context.CancelFunc
-	done     chan struct{}
-	waiters  map[string]progressWaiter
-	// unsupported is set once opening the connection fails outright (e.g. a
-	// 404/405, which some upstreams legitimately return for this endpoint
-	// per spec). It stops every later acquire from re-attempting a doomed
-	// connection on every single streaming call; it resets naturally the
-	// next time refCount drops to zero and this entry is evicted.
-	unsupported bool
-}
-
-// acquireStandaloneStream returns the shared standaloneStream for upstream,
-// creating it and starting its reader goroutine if this is the first
-// waiter. Callers must pair this with exactly one releaseStandaloneStream.
-func (c *Client) acquireStandaloneStream(upstream Upstream) *standaloneStream {
-	c.standaloneMu.Lock()
-	s := c.standaloneStreams[upstream.Name]
-	if s == nil {
-		s = &standaloneStream{waiters: make(map[string]progressWaiter)}
-		c.standaloneStreams[upstream.Name] = s
-	}
-	c.standaloneMu.Unlock()
-
-	s.mu.Lock()
-	s.refCount++
-	start := s.refCount == 1 && !s.unsupported
-	if start {
-		streamCtx, cancel := context.WithCancel(context.Background())
-		s.cancel = cancel
-		s.done = make(chan struct{})
-		go c.runStandaloneStream(streamCtx, upstream, s)
-	}
-	s.mu.Unlock()
-	return s
-}
-
-// releaseStandaloneStream drops one reference acquired via
-// acquireStandaloneStream. Once the last reference is gone, it cancels the
-// reader goroutine, waits for it to fully exit, and evicts the entry so a
-// future acquire opens a fresh connection (picking up, e.g., a session that
-// was reinitialized in the meantime).
-func (c *Client) releaseStandaloneStream(upstream Upstream, s *standaloneStream) {
-	s.mu.Lock()
-	s.refCount--
-	last := s.refCount <= 0
-	var cancel context.CancelFunc
-	var done chan struct{}
-	if last {
-		cancel = s.cancel
-		done = s.done
-		s.cancel = nil
-		s.done = nil
-	}
-	s.mu.Unlock()
-	if cancel != nil {
-		cancel()
-		<-done
-	}
-	if last {
-		c.standaloneMu.Lock()
-		if c.standaloneStreams[upstream.Name] == s {
-			delete(c.standaloneStreams, upstream.Name)
-		}
-		c.standaloneMu.Unlock()
-	}
-}
-
-func (s *standaloneStream) registerWaiter(token string, w progressWaiter) {
-	s.mu.Lock()
-	s.waiters[token] = w
-	s.mu.Unlock()
-}
-
-func (s *standaloneStream) unregisterWaiter(token string) {
-	s.mu.Lock()
-	delete(s.waiters, token)
-	s.mu.Unlock()
-}
-
-// openStandaloneGET opens the standalone SSE stream: a bare GET carrying the
-// session's headers, no Last-Event-ID (see standaloneStream doc comment).
-// Mirrors resumeSSEStream's header handling.
-func (c *Client) openStandaloneGET(ctx context.Context, upstream Upstream) (*http.Response, error) {
-	endpoint := strings.TrimRight(upstream.BaseURL, "/")
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Accept", "text/event-stream")
-	if protocol := c.getSessionProtocol(upstream.Name); protocol != "" {
-		req.Header.Set("MCP-Protocol-Version", protocol)
-	}
-	if sessionID := c.getSession(upstream.Name); sessionID != "" {
-		req.Header.Set("Mcp-Session-Id", sessionID)
-	}
-	applyAuthHeaders(req, upstream)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode >= http.StatusBadRequest {
-		defer resp.Body.Close()
-		body, _ := io.ReadAll(io.LimitReader(resp.Body, 64*1024))
-		return nil, fmt.Errorf("upstream %q standalone stream returned HTTP %d: %s", upstream.Name, resp.StatusCode, extractErrorDetail(body))
-	}
-	if !strings.Contains(strings.ToLower(resp.Header.Get("Content-Type")), "text/event-stream") {
-		defer resp.Body.Close()
-		return nil, fmt.Errorf("upstream %q standalone stream returned content type %q, want text/event-stream", upstream.Name, resp.Header.Get("Content-Type"))
-	}
-	return resp, nil
-}
-
-func (c *Client) runStandaloneStream(ctx context.Context, upstream Upstream, s *standaloneStream) {
-	defer close(s.done)
-	resp, err := c.openStandaloneGET(ctx, upstream)
-	if err != nil {
-		c.debugf("standalone stream unavailable server=%s err=%v", upstream.Name, err)
-		s.mu.Lock()
-		s.unsupported = true
-		s.mu.Unlock()
-		return
-	}
-	defer resp.Body.Close()
-	reader := newSSEEventReader(resp.Body)
-	for {
-		evt, err := reader.NextEvent()
-		if err != nil {
-			return
-		}
-		if !evt.HasData {
-			continue
-		}
-		c.routeStandaloneEvent(s, evt.Data)
-	}
-}
-
-// routeStandaloneEvent attributes one standalone-stream message to whichever
-// registered call it belongs to. Progress notifications carry the token
-// Atryum minted for that call (see rewriteProgressToken) in
-// params.progressToken, giving an unambiguous match. Anything else (e.g. a
-// logging notification) carries no per-call correlator at all; it is
-// delivered only when exactly one call is currently waiting on this stream,
-// since there is no way to attribute it correctly when several calls are
-// in flight concurrently — and silently guessing wrong would leak one
-// caller's message to another.
-func (c *Client) routeStandaloneEvent(s *standaloneStream, payload []byte) {
-	var message map[string]json.RawMessage
-	if err := json.Unmarshal(payload, &message); err != nil {
-		return
-	}
-	if _, hasMethod := message["method"]; !hasMethod {
-		return
-	}
-	wireToken, hasToken := extractProgressToken(message)
-
-	s.mu.Lock()
-	var waiter progressWaiter
-	var ok bool
-	if hasToken {
-		waiter, ok = s.waiters[wireToken]
-	} else if len(s.waiters) == 1 {
-		for _, w := range s.waiters {
-			waiter, ok = w, true
-		}
-	}
-	s.mu.Unlock()
-	if !ok {
-		return
-	}
-
-	// Handed off to the matching call's own goroutine via its channel — see
-	// progressWaiter for why this indirection matters. callSink.Event (on
-	// the receiving end) restores the caller's original progressToken
-	// itself (matching on its own wireToken), so the raw payload is sent
-	// through unmodified here.
-	select {
-	case waiter.events <- StreamEvent{Data: payload}:
-	default:
-		// Buffer full, or the receiving call already stopped draining it —
-		// drop rather than block this shared reader goroutine, which also
-		// serves every other call currently sharing this connection.
-	}
-}
-
-// extractProgressToken reads params.progressToken from an already-decoded
-// JSON-RPC message, normalizing it to a bare string for map lookup
-// regardless of whether the upstream echoed it back as a JSON string or a
-// number.
-func extractProgressToken(message map[string]json.RawMessage) (string, bool) {
-	paramsRaw, ok := message["params"]
-	if !ok {
-		return "", false
-	}
-	var params struct {
-		ProgressToken json.RawMessage `json:"progressToken"`
-	}
-	if err := json.Unmarshal(paramsRaw, &params); err != nil || len(params.ProgressToken) == 0 {
-		return "", false
-	}
-	return strings.Trim(string(params.ProgressToken), `"`), true
-}
-
-// rewriteProgressTokenInPayload replaces params.progressToken in an
-// already-wire-formatted JSON-RPC message with originalToken, restoring the
-// value the caller actually supplied before relaying the message onward.
-func rewriteProgressTokenInPayload(payload []byte, originalToken any) ([]byte, error) {
-	var generic map[string]any
-	if err := json.Unmarshal(payload, &generic); err != nil {
-		return nil, err
-	}
-	params, ok := generic["params"].(map[string]any)
-	if !ok {
-		return nil, fmt.Errorf("message has no params object")
-	}
-	params["progressToken"] = originalToken
-	generic["params"] = params
-	return json.Marshal(generic)
-}
-
-// rewriteProgressToken replaces meta's progressToken, if any, with a value
-// unique to this specific call, returning the rewritten meta, that wire
-// token, and the caller's original token. Atryum multiplexes every
-// downstream caller of a given upstream onto one shared session, so two
-// unrelated concurrent calls could independently pick the same
-// caller-supplied progressToken; rewriting to a per-call value here is what
-// lets routeStandaloneEvent attribute a notification to the right call
-// instead of risking a cross-call delivery.
-func (c *Client) rewriteProgressToken(meta map[string]any) (rewritten map[string]any, wireToken string, original any, ok bool) {
-	if meta == nil {
-		return meta, "", nil, false
-	}
-	original, ok = meta["progressToken"]
-	if !ok {
-		return meta, "", nil, false
-	}
-	wireToken = fmt.Sprintf("atryum-pt-%d", c.nextID.Add(1))
-	rewritten = make(map[string]any, len(meta))
-	for k, v := range meta {
-		rewritten[k] = v
-	}
-	rewritten["progressToken"] = wireToken
-	return rewritten, wireToken, original, true
-}
-
-// callSink wraps the caller's sink for one streaming call that requested
-// progress tracking, restoring the caller's original progressToken in
-// place of the wire-level token Atryum minted (see rewriteProgressToken) on
-// every Event call — regardless of whether relaySSEToolCall read the
-// message from the call's own POST response or from the standalone
-// stream's per-call channel (see progressWaiter). Some upstreams echo a
-// call's progress notifications on the tools/call POST response itself
-// rather than the standalone stream — that's the more spec-typical case,
-// in fact — so the restore can't live only in the standalone-delivery
-// path, or the agent would see Atryum's internal token leak through there.
-//
-// relaySSEToolCall drains both sources from a single goroutine (see
-// postStreamPump), so, unlike an earlier version of this type, Event and
-// StreamStarted need no guard against concurrent calls.
-type callSink struct {
-	inner         StreamSink
-	wireToken     string
-	originalToken any
-}
-
-func newCallSink(inner StreamSink, wireToken string, originalToken any) *callSink {
-	return &callSink{inner: inner, wireToken: wireToken, originalToken: originalToken}
-}
-
-func (s *callSink) StreamStarted() {
-	s.inner.StreamStarted()
-}
-
-func (s *callSink) Event(evt StreamEvent) error {
-	var message map[string]json.RawMessage
-	if err := json.Unmarshal(evt.Data, &message); err == nil {
-		if token, ok := extractProgressToken(message); ok && token == s.wireToken {
-			if rewritten, err := rewriteProgressTokenInPayload(evt.Data, s.originalToken); err == nil {
-				evt.Data = rewritten
-			}
-		}
-	}
-	return s.inner.Event(evt)
-}
-
-// resumeSSEStream continues a server-closed Streamable HTTP response. The
-// MCP transport specifies a GET to the same endpoint carrying Last-Event-ID;
-// session, protocol, and authentication headers must match the original
-// connection so the upstream can locate the pending request.
-func (c *Client) resumeSSEStream(ctx context.Context, upstream Upstream, lastEventID string) (*http.Response, error) {
-	endpoint := strings.TrimRight(upstream.BaseURL, "/")
-	req, err := http.NewRequestWithContext(ctx, http.MethodGet, endpoint, nil)
-	if err != nil {
-		return nil, err
-	}
-	req.Header.Set("Accept", "text/event-stream")
-	req.Header.Set("Last-Event-ID", lastEventID)
-	if protocol := c.getSessionProtocol(upstream.Name); protocol != "" {
-		req.Header.Set("MCP-Protocol-Version", protocol)
-	}
-	if sessionID := c.getSession(upstream.Name); sessionID != "" {
-		req.Header.Set("Mcp-Session-Id", sessionID)
-	}
-	applyAuthHeaders(req, upstream)
-	resp, err := c.httpClient.Do(req)
-	if err != nil {
-		return nil, err
-	}
-	if resp.StatusCode >= http.StatusBadRequest {
-		defer resp.Body.Close()
-		body, _ := io.ReadAll(io.LimitReader(resp.Body, 64*1024))
-		return nil, fmt.Errorf("upstream %q resume failed with HTTP %d: %s", upstream.Name, resp.StatusCode, extractErrorDetail(body))
-	}
-	if !strings.Contains(strings.ToLower(resp.Header.Get("Content-Type")), "text/event-stream") {
-		defer resp.Body.Close()
-		return nil, fmt.Errorf("upstream %q resume returned content type %q, want text/event-stream", upstream.Name, resp.Header.Get("Content-Type"))
-	}
-	if newSession := strings.TrimSpace(resp.Header.Get("Mcp-Session-Id")); newSession != "" {
-		protocol := c.getSessionProtocol(upstream.Name)
-		c.setSession(upstream.Name, newSession, protocol)
-	}
-	return resp, nil
-}
-
-// runSessionInitBounded runs fn (a session initialize/reinitialize step)
-// under opts.HeaderTimeout. The session-init POSTs happen before the
-// streaming call proper, so doHTTPToolCallStream's own header timeout never
-// covers them; and in streaming mode the caller's ctx carries no deadline
-// (the fixed request timeout is deliberately not applied — that's the whole
-// point of the streaming timeout scheme). Without this bound, an upstream
-// with no per-server timeout_seconds configured that hangs during
-// initialize would block the call indefinitely.
-func runSessionInitBounded(ctx context.Context, upstream Upstream, opts StreamOptions, fn func(context.Context) error) error {
-	initCtx := ctx
-	if opts.HeaderTimeout > 0 {
-		var cancel context.CancelFunc
-		initCtx, cancel = context.WithTimeout(ctx, opts.HeaderTimeout)
-		defer cancel()
-	}
-	err := fn(initCtx)
-	if err != nil && initCtx.Err() != nil && ctx.Err() == nil {
-		// The bound we imposed fired (not the caller's own ctx): surface it
-		// as the same typed timeout the rest of the streaming path uses.
-		return fmt.Errorf("upstream %q: timed out initializing session before streaming call: %w", upstream.Name, ErrStreamTimeout)
-	}
-	return err
-}
-
-func (c *Client) invokeHTTPStream(ctx context.Context, upstream Upstream, tool string, input map[string]any, requestID *string, meta map[string]any, sink StreamSink, opts StreamOptions) (InvokeResult, error) {
-	if err := runSessionInitBounded(ctx, upstream, opts, func(initCtx context.Context) error {
-		return c.ensureHTTPSession(initCtx, upstream)
-	}); err != nil {
-		return InvokeResult{}, err
-	}
-
-	merged := mergeRequestMeta(meta, requestID)
-	// effectiveSink is what actually gets passed to doHTTPToolCallStream.
-	// When this call requested progress tracking, it becomes a *callSink
-	// that restores the caller's original progressToken on every Event
-	// call, regardless of which of the two upstream channels (this call's
-	// own POST response, or the standalone stream via progressCh) the
-	// underlying message arrived on.
-	effectiveSink := sink
-	var progressCh chan StreamEvent
-	if rewritten, wireToken, original, ok := c.rewriteProgressToken(merged); ok {
-		merged = rewritten
-		effectiveSink = newCallSink(sink, wireToken, original)
-		progressCh = make(chan StreamEvent, standaloneWaiterEventBuffer)
-		standalone := c.acquireStandaloneStream(upstream)
-		standalone.registerWaiter(wireToken, progressWaiter{events: progressCh})
-		defer func() {
-			c.releaseStandaloneStream(upstream, standalone)
-			// Deliberately not unregistered synchronously here: this call's
-			// own POST-response stream and the shared standalone connection
-			// are two independent connections read by two independent
-			// goroutines, with no ordering guarantee between them. A
-			// notification for this exact call can still be in flight on the
-			// standalone connection at the moment this call's own terminal
-			// response arrives — removing the waiter immediately risks the
-			// reader goroutine finding nothing for a notification that was
-			// legitimately on its way, silently dropping it. Wire tokens are
-			// never reused (always a fresh atomic counter value), so nothing
-			// is unsafe about the waiter lingering a little longer; delaying
-			// the removal trades a small, bounded amount of memory for
-			// closing that window.
-			time.AfterFunc(standaloneWaiterGracePeriod, func() {
-				standalone.unregisterWaiter(wireToken)
-			})
-		}()
-	}
-
-	body, err := marshalToolCallEnvelopeWithMeta(tool, input, merged)
-	if err != nil {
-		return InvokeResult{}, err
-	}
-
-	outcome, err := c.doHTTPToolCallStream(ctx, upstream, body, effectiveSink, progressCh, opts)
-	if err != nil {
-		return InvokeResult{}, err
-	}
-	if outcome.missingSession {
-		if outcome.eventsRelayed > 0 {
-			return InvokeResult{}, fmt.Errorf("upstream %q reported a missing session after the stream had already relayed %d event(s): %w", upstream.Name, outcome.eventsRelayed, ErrStreamSessionRetryRefused)
-		}
-		c.debugf("upstream http tools.call stream missing session server=%s session=%q", upstream.Name, outcome.sessionID)
-		if retryErr := runSessionInitBounded(ctx, upstream, opts, func(initCtx context.Context) error {
-			return c.reinitializeRequiredHTTPSession(initCtx, upstream, outcome.sessionID)
-		}); retryErr != nil {
-			return InvokeResult{}, retryErr
-		}
-		outcome, err = c.doHTTPToolCallStream(ctx, upstream, body, effectiveSink, progressCh, opts)
-		if err != nil {
-			return InvokeResult{}, err
-		}
-		if outcome.missingSession {
-			return InvokeResult{}, fmt.Errorf("upstream %q rejected session after reinitialize", upstream.Name)
-		}
-	}
-	c.debugf("upstream http tools.call stream server=%s status=%d failed=%t events_relayed=%d", upstream.Name, outcome.invoke.StatusCode, outcome.invoke.Failed, outcome.eventsRelayed)
-	return outcome.invoke, nil
-}
-
-// InvokeStream behaves like Invoke while also relaying intermediate JSON-RPC
-// messages to sink as they arrive. HTTP upstreams select streaming with an
-// SSE response, so StreamStarted fires even when that SSE response contains
-// only its terminal message. Stdio has no equivalent transport signal, so its
-// sink starts only when an intermediate message arrives. A nil sink always
-// uses Invoke's buffered path.
-func (c *Client) InvokeStream(ctx context.Context, upstream Upstream, tool string, input map[string]any, requestID *string, meta map[string]any, sink StreamSink, opts StreamOptions) (InvokeResult, error) {
-	switch upstream.Mode {
-	case UpstreamModeStdio:
-		if sink == nil {
-			return c.Invoke(ctx, upstream, tool, input, requestID, meta)
-		}
-		started := time.Now()
-		defer func() {
-			c.debugf("upstream invoke-stream transport=%s server=%s tool=%s duration_ms=%d", upstream.Mode, upstream.Name, tool, time.Since(started).Milliseconds())
-		}()
-		return c.invokeStdioStream(ctx, upstream, tool, input, requestID, meta, sink, opts)
-	case UpstreamModeHTTP, "":
-		if sink == nil {
-			return c.Invoke(ctx, upstream, tool, input, requestID, meta)
-		}
-		started := time.Now()
-		defer func() {
-			c.debugf("upstream invoke-stream transport=%s server=%s tool=%s duration_ms=%d", upstream.Mode, upstream.Name, tool, time.Since(started).Milliseconds())
-		}()
-		return c.invokeHTTPStream(ctx, upstream, tool, input, requestID, meta, sink, opts)
-	default:
-		return InvokeResult{}, fmt.Errorf("unsupported upstream mode %q", upstream.Mode)
-	}
-}
-
 func (c *Client) listToolsHTTP(ctx context.Context, upstream Upstream) ([]Tool, error) {
 	if err := c.ensureHTTPSession(ctx, upstream); err != nil {
 		return nil, err
@@ -2306,155 +1191,6 @@ func (c *Client) invokeStdio(ctx context.Context, upstream Upstream, tool string
 	return InvokeResult{StatusCode: http.StatusOK, Body: body, Failed: looksLikeToolError(body)}, nil
 }
 
-// invokeStdioStream is invokeStdio's live-relay counterpart. Unlike HTTP,
-// stdio has no header phase or Content-Type to signal in advance whether
-// the upstream will emit anything beyond its terminal response — every
-// stdio reply is the same newline-delimited JSON-RPC framing regardless.
-// So instead of a transport-level signal, StreamStarted fires lazily, the
-// same way the HTTP path already does for its own edge case (see
-// relaySSEToolCall): only right before the first actual notification or
-// server-request is relayed. A call that produces nothing but its terminal
-// response never touches sink at all, staying identical to invokeStdio.
-func (c *Client) invokeStdioStream(ctx context.Context, upstream Upstream, tool string, input map[string]any, requestID *string, meta map[string]any, sink StreamSink, opts StreamOptions) (InvokeResult, error) {
-	if upstream.Command == "" {
-		return InvokeResult{}, fmt.Errorf("stdio upstream %q missing command", upstream.Name)
-	}
-	guard := newCallTimeoutGuard(ctx)
-	defer guard.stop()
-
-	cmd := exec.CommandContext(guard.ctx, upstream.Command, upstream.Args...)
-	cmd.Env = os.Environ()
-	for k, v := range upstream.Env {
-		cmd.Env = append(cmd.Env, k+"="+v)
-	}
-	configureStdioProcessGroup(cmd)
-	stdin, err := cmd.StdinPipe()
-	if err != nil {
-		return InvokeResult{}, err
-	}
-	stdout, err := cmd.StdoutPipe()
-	if err != nil {
-		return InvokeResult{}, err
-	}
-	stderr := newBoundedBuffer(stdioStderrCap)
-	cmd.Stderr = stderr
-	if err := cmd.Start(); err != nil {
-		return InvokeResult{}, err
-	}
-	defer func() {
-		// guard.stop() MUST run before cmd.Wait(): stopping cancels the
-		// guard context, which triggers the process-group kill, which is
-		// what makes Wait return. Deferring these separately would run
-		// them in LIFO order — Wait before stop — and a stdio server that
-		// keeps running after answering (normal for long-lived servers)
-		// or that ignores stdin-close would then block Wait forever on
-		// every return path where no timeout had fired (sink error, or a
-		// successfully received terminal response). The process is
-		// per-call and disposable, so killing it once we have our answer
-		// (or have given up) is the correct lifecycle.
-		guard.stop()
-		_ = stdin.Close()
-		_ = cmd.Wait()
-	}()
-
-	reader := bufio.NewReader(stdout)
-	// The initialize handshake gets the same header-phase bound the HTTP
-	// path applies before its response headers arrive: without it, a
-	// subprocess that starts but never answers initialize would block
-	// readRPC with no bound of its own (only the caller's ctx).
-	guard.armHeaderTimeout(opts.HeaderTimeout)
-	initID := c.nextRPCID()
-	if err := writeRPC(stdin, initID, "initialize", map[string]any{
-		"protocolVersion": DefaultMCPProtocolVersion,
-		"clientInfo":      map[string]any{"name": "atryum", "version": version.Version},
-		"capabilities":    map[string]any{},
-	}); err != nil {
-		return InvokeResult{}, err
-	}
-	if _, err := readRPC(reader, rpcIDMessage(initID)); err != nil {
-		if reason := guard.reason(); reason != "" {
-			return InvokeResult{}, fmt.Errorf("upstream %q: %s during stdio initialize: %w", upstream.Name, reason, ErrStreamTimeout)
-		}
-		if stderr.Len() > 0 {
-			return InvokeResult{}, fmt.Errorf("stdio upstream error: %s", strings.TrimSpace(stderr.String()))
-		}
-		return InvokeResult{}, err
-	}
-	guard.disarmHeaderTimeout()
-	_ = writeRPC(stdin, c.nextRPCID(), "notifications/initialized", map[string]any{})
-	callParams := map[string]any{"name": tool, "arguments": input}
-	if merged := mergeRequestMeta(meta, requestID); merged != nil {
-		callParams["_meta"] = merged
-	}
-	callID := c.nextRPCID()
-	if err := writeRPC(stdin, callID, "tools/call", callParams); err != nil {
-		return InvokeResult{}, err
-	}
-
-	guard.armBodyTimeouts(opts.IdleTimeout, opts.MaxDuration)
-	return c.relayStdioToolCall(reader, sink, guard, upstream, callID, stderr)
-}
-
-// relayStdioToolCall reads reader's newline-delimited JSON-RPC messages,
-// relaying every intermediate (non-terminal) message to sink as it arrives,
-// and returns once the terminal response for callID is read.
-func (c *Client) relayStdioToolCall(reader *bufio.Reader, sink StreamSink, guard *callTimeoutGuard, upstream Upstream, callID int64, stderr *boundedBuffer) (InvokeResult, error) {
-	expectedID := rpcIDMessage(callID)
-	started := false
-	ensureStarted := func() {
-		if !started {
-			started = true
-			sink.StreamStarted()
-		}
-	}
-	for {
-		line, err := reader.ReadBytes('\n')
-		if err != nil {
-			if reason := guard.reason(); reason != "" {
-				return InvokeResult{}, fmt.Errorf("upstream %q %s: %w", upstream.Name, reason, ErrStreamTimeout)
-			}
-			if stderr.Len() > 0 {
-				return InvokeResult{}, fmt.Errorf("stdio upstream error: %s", strings.TrimSpace(stderr.String()))
-			}
-			return InvokeResult{}, err
-		}
-		line = bytes.TrimSpace(line)
-		if len(line) == 0 {
-			continue
-		}
-		guard.resetIdle()
-
-		switch classifyRPCMessage(line, expectedID) {
-		case rpcMessageTerminalResponse:
-			var resp rpcResponse
-			if err := json.Unmarshal(line, &resp); err != nil {
-				continue
-			}
-			if len(resp.Error) > 0 && string(resp.Error) != "null" {
-				return InvokeResult{StatusCode: http.StatusBadGateway, Body: resp.Error, Failed: true}, nil
-			}
-			body := resp.Result
-			if len(body) == 0 {
-				body = []byte(`{"ok":true}`)
-			}
-			return InvokeResult{StatusCode: http.StatusOK, Body: body, Failed: looksLikeToolError(body)}, nil
-		case rpcMessageServerRequest:
-			ensureStarted()
-			if err := sink.Event(StreamEvent{Data: line, ServerRequest: true}); err != nil {
-				return InvokeResult{}, err
-			}
-		case rpcMessageNotification:
-			ensureStarted()
-			if err := sink.Event(StreamEvent{Data: line}); err != nil {
-				return InvokeResult{}, err
-			}
-		default:
-			// Unparseable line, or a response to some other id. Not ours
-			// to interpret; ignore and keep reading.
-		}
-	}
-}
-
 func (c *Client) listToolsStdio(ctx context.Context, upstream Upstream) ([]Tool, error) {
 	if upstream.Command == "" {
 		return nil, fmt.Errorf("stdio upstream %q missing command", upstream.Name)
@@ -2715,143 +1451,6 @@ func (c *Client) testStdio(ctx context.Context, upstream Upstream) ConnectionTes
 		return ConnectionTestResult{Ok: false, Message: message, ConnectionStatus: ConnectionStatusNeedsAttention, AuthStatus: authStatus, ReauthNeeded: reauth, LastCheckOK: false, LastErrorSummary: &message, ActionRequired: action}
 	}
 	return ConnectionTestResult{Ok: true, Message: "stdio initialize ok", ConnectionStatus: ConnectionStatusReady, AuthStatus: AuthStatusReady, ReauthNeeded: false, LastCheckOK: true}
-}
-
-// sseEventReader incrementally parses a Server-Sent Events body, returning
-// one joined "data:" payload per event via Next. It is the shared parser
-// behind both the buffered SSE consumers (extractSSEJSONRPCResponse, used by
-// tools/list, initialize, and the default forward path) and the incremental
-// streaming relay (relaySSEToolCall) — one parser, two ways of consuming it.
-type sseEventReader struct {
-	scanner   *bufio.Scanner
-	dataLines []string
-	eventID   string
-	retry     time.Duration
-	hasData   bool
-	hasID     bool
-	hasRetry  bool
-}
-
-type sseWireEvent struct {
-	Data     []byte
-	ID       string
-	Retry    time.Duration
-	HasData  bool
-	HasID    bool
-	HasRetry bool
-}
-
-func newSSEEventReader(r io.Reader) *sseEventReader {
-	scanner := bufio.NewScanner(r)
-	scanner.Buffer(make([]byte, 1024*1024), 4*1024*1024)
-	return &sseEventReader{scanner: scanner}
-}
-
-// NextEvent returns one complete SSE event, including the id/retry fields
-// needed to resume a Streamable HTTP response after the upstream closes it.
-func (r *sseEventReader) NextEvent() (sseWireEvent, error) {
-	for r.scanner.Scan() {
-		line := r.scanner.Text()
-		if line == "" {
-			if !r.hasData && !r.hasID && !r.hasRetry {
-				continue
-			}
-			return r.takeEvent(), nil
-		}
-		if strings.HasPrefix(line, ":") {
-			continue
-		}
-		field, value, ok := strings.Cut(line, ":")
-		if !ok {
-			field = line
-			value = ""
-		} else if strings.HasPrefix(value, " ") {
-			value = strings.TrimPrefix(value, " ")
-		}
-		switch field {
-		case "data":
-			r.dataLines = append(r.dataLines, value)
-			r.hasData = true
-		case "id":
-			// The SSE specification ignores id values containing NUL.
-			if !strings.ContainsRune(value, '\x00') {
-				r.eventID = value
-				r.hasID = true
-			}
-		case "retry":
-			millis, err := strconv.ParseInt(value, 10, 64)
-			if err == nil && millis >= 0 {
-				const maxRetryMillis = int64((time.Duration(1<<63 - 1)) / time.Millisecond)
-				if millis > maxRetryMillis {
-					r.retry = time.Duration(1<<63 - 1)
-				} else {
-					r.retry = time.Duration(millis) * time.Millisecond
-				}
-				r.hasRetry = true
-			}
-		}
-	}
-	if err := r.scanner.Err(); err != nil {
-		return sseWireEvent{}, err
-	}
-	if r.hasData || r.hasID || r.hasRetry {
-		return r.takeEvent(), nil
-	}
-	return sseWireEvent{}, io.EOF
-}
-
-func (r *sseEventReader) takeEvent() sseWireEvent {
-	evt := sseWireEvent{
-		Data:     []byte(strings.Join(r.dataLines, "\n")),
-		ID:       r.eventID,
-		Retry:    r.retry,
-		HasData:  r.hasData,
-		HasID:    r.hasID,
-		HasRetry: r.hasRetry,
-	}
-	r.dataLines = nil
-	r.eventID = ""
-	r.retry = 0
-	r.hasData = false
-	r.hasID = false
-	r.hasRetry = false
-	return evt
-}
-
-// Next is the payload-only view used by buffered consumers. Control-only
-// events (id/retry with no data) are skipped because they carry no JSON-RPC
-// message for those callers to decode.
-func (r *sseEventReader) Next() ([]byte, error) {
-	for {
-		evt, err := r.NextEvent()
-		if err != nil {
-			return nil, err
-		}
-		if evt.HasData {
-			return evt.Data, nil
-		}
-	}
-}
-
-// extractSSEJSONRPCResponse scans an SSE body for the one event that is
-// either the response matching expectedID or the null-id error JSON-RPC
-// uses when a server can't identify which request an error belongs to,
-// skipping everything else (notifications, unrelated responses).
-func extractSSEJSONRPCResponse(r io.Reader, expectedID json.RawMessage) ([]byte, error) {
-	reader := newSSEEventReader(r)
-	for {
-		payload, err := reader.Next()
-		if err == io.EOF {
-			return nil, fmt.Errorf("no JSON-RPC response in SSE stream")
-		}
-		if err != nil {
-			return nil, err
-		}
-		match := classifyJSONRPCResponsePayload(payload, expectedID)
-		if match == jsonRPCResponseIDMatch || match == jsonRPCResponseNullIDError {
-			return payload, nil
-		}
-	}
 }
 
 type jsonRPCResponseMatch int
